requirements.txt
# Django Framework
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Google Maps
googlemaps==4.10.0

# Email & Reports
reportlab==4.0.9
openpyxl==3.1.2
pillow==10.2.0

# Security & Rate Limiting
django-ratelimit==4.1.0

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
django-stubs==4.2.7
faker==22.4.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Environment Variables
python-decouple==3.8